Added integration testing for captureFetch
