+++ packages/core/lib/aws-xray.d.ts
@@ -44,8 +44,6 @@ export { captureAWSClient as captureAWSv3Client } from './patchers/aws3_p';
 
 export { captureHTTPs, captureHTTPsGlobal } from './patchers/http_p';
 
-export { captureFetch } from './patchers/fetch_p';
-
 export { capturePromise } from './patchers/promise_p';
 
 export { utils };
