+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aws-xray-sdk-core",
-  "version": "3.4.1",
+  "version": "3.5.0",
   "description": "AWS X-Ray SDK for Javascript",
   "author": "Amazon Web Services",
   "contributors": [
@@ -23,16 +23,12 @@
   },
   "//": "@types/cls-hooked is exposed in API so must be in dependencies, not devDependencies",
   "dependencies": {
-    "@aws-sdk/service-error-classification": "^3.4.1",
     "@aws-sdk/types": "^3.4.1",
+    "@aws-sdk/service-error-classification": "^3.4.1",
     "@types/cls-hooked": "^4.3.3",
-    "@types/node-fetch": "^2.6.3",
     "atomic-batcher": "^1.0.2",
     "cls-hooked": "^4.2.2",
-    "mocha": "^10.2.0",
-    "semver": "^5.3.0",
-    "tsd": "^0.28.1",
-    "typescript": "^5.0.4"
+    "semver": "^7.3.8"
   },
   "scripts": {
     "prepare": "npm run compile",
@@ -42,7 +38,6 @@
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "test": "npm run compile && mocha --recursive ./dist/test/ -R spec && tsd && mocha --recursive ./dist/test_async/ -R spec",
-    "test-grep": "npm run compile && mocha --recursive ./dist/test/ -R spec --fgrep $TEST && tsd && mocha --recursive ./dist/test_async/ -R spec --fgrep $TEST",
     "test-d": "tsd",
     "test-async": "npm run compile && mocha --recursive ./dist/test_async/ -R spec",
     "clean": "rm -rf dist && rm -rf node_modules",
