+++ sdk_contrib/fetch/package.json
@@ -45,5 +45,9 @@
     "node-fetch": "^2.6.11",
     "ts-expect": "^1.3.0",
     "ts-node": "^10.9.1"
+  },
+  "dependencies": {
+    "mocha": "^10.2.0",
+    "tsd": "^0.28.1"
   }
 }
